+++ src/main/java/org/apache/tomcat/jakartaee/Migration.java
@@ -32,8 +32,11 @@ import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import java.util.zip.CRC32;
+import java.util.zip.ZipEntry;
 import java.util.zip.ZipException;
 
+import org.apache.commons.compress.archivers.ArchiveEntry;
 import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
 import org.apache.commons.compress.archivers.zip.ZipArchiveInputStream;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
@@ -211,8 +214,8 @@ public class Migration {
 
 
     private void migrateArchiveStreaming(InputStream src, OutputStream dest) throws IOException {
-        try (ZipArchiveInputStream srcZipStream = new ZipArchiveInputStream(new CloseShieldInputStream(src));
-                ZipArchiveOutputStream destZipStream = new ZipArchiveOutputStream(new CloseShieldOutputStream(dest))) {
+        try (ZipArchiveInputStream srcZipStream = new ZipArchiveInputStream(CloseShieldInputStream.wrap(src));
+                CrcZipArchiveOutputStream destZipStream = new CrcZipArchiveOutputStream(CloseShieldOutputStream.wrap(dest))) {
             ZipArchiveEntry srcZipEntry;
             while ((srcZipEntry = srcZipStream.getNextZipEntry()) != null) {
                 String srcName = srcZipEntry.getName();
@@ -221,7 +224,7 @@ public class Migration {
                     continue;
                 }
                 String destName = profile.convert(srcName);
-                RenamableZipArchiveEntry destZipEntry = new RenamableZipArchiveEntry(srcZipEntry);
+                MigrationZipArchiveEntry destZipEntry = new MigrationZipArchiveEntry(srcZipEntry, false);
                 destZipEntry.setName(destName);
                 destZipStream.putArchiveEntry(destZipEntry);
                 migrateStream(srcName, srcZipStream, destZipStream);
@@ -251,7 +254,7 @@ public class Migration {
                     continue;
                 }
                 String destName = profile.convert(srcName);
-                RenamableZipArchiveEntry destZipEntry = new RenamableZipArchiveEntry(srcZipEntry);
+                MigrationZipArchiveEntry destZipEntry = new MigrationZipArchiveEntry(srcZipEntry, true);
                 destZipEntry.setName(destName);
                 destZipStream.putArchiveEntry(destZipEntry);
                 migrateStream(srcName, srcZipFile.getInputStream(srcZipEntry), destZipStream);
@@ -316,15 +319,73 @@ public class Migration {
         return false;
     }
 
-    private static class RenamableZipArchiveEntry extends ZipArchiveEntry {
-
-        public RenamableZipArchiveEntry(ZipArchiveEntry entry) throws ZipException {
+    private static class MigrationZipArchiveEntry extends ZipArchiveEntry {
+        protected final CRC32 crc = new CRC32();
+        protected long size = 0;
+        protected boolean needResetCrc;
+        public MigrationZipArchiveEntry(ZipArchiveEntry entry, boolean inMemory) throws ZipException {
             super(entry);
+            // No recalculation required, when in memory mode and not of type SORTED
+            needResetCrc = !inMemory && entry.getMethod() == ZipEntry.STORED;
+        }
+
+        @Override
+        public long getSize() {
+            return needResetCrc ? size : super.getSize();
+        }
+
+        @Override
+        public long getCrc() {
+            return needResetCrc ? crc.getValue() : super.getCrc();
+        }
+
+        public void update(byte[] b, int offset, int length) {
+            if (needResetCrc) {
+                crc.update(b, offset, length);
+                size += length;
+            }
         }
 
         @Override
         public void setName(String name) {
             super.setName(name);
+        }
+    }
+
+    private static class CrcZipArchiveOutputStream extends ZipArchiveOutputStream {
+        private MigrationZipArchiveEntry current;
+        private CrcZipArchiveOutputStream(OutputStream out) {
+            super(out);
+        }
+
+        @Override
+        public void write(byte[] b, int offset, int length) throws IOException {
+            super.write(b, offset, length);
+            update(b, offset, length);
+        }
+
+        @Override
+        public void putArchiveEntry(ArchiveEntry archiveEntry) throws IOException {
+            if (archiveEntry instanceof MigrationZipArchiveEntry) {
+                current = (MigrationZipArchiveEntry) archiveEntry;
+            }
+            super.putArchiveEntry(archiveEntry);
+        }
+
+        @Override
+        public void closeArchiveEntry() throws IOException {
+            reset();
+            super.closeArchiveEntry();
+        }
+
+        private void reset() {
+            current = null;
+        }
+
+        private void update(byte[] b, int offset, int length) {
+            if (current != null) {
+                current.update(b, offset, length);
+            }
         }
     }
 }